requirements.txt
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
beautifulsoup4==4.13.4
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
flake8==7.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.16.0
mypy-extensions==1.1.0
openai==1.86.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycodestyle==2.13.0
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydentic==0.0.1.dev3
pyflakes==3.3.2
pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-stdnum==2.1
pyyaml==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
